aco hdf5 v5.5 - less storage

+++ src/ant_colony/aco_grad_h5v5.py
@@ -74,7 +74,7 @@ class GradACO:
                     d[i][j] += np.sum(temp_bin)
         # print(d)
         grp_name = 'dataset/' + self.d_set.step_name + '/d_matrix'
-        self.d_set.add_h5_dataset(grp_name, d)
+        self.d_set.add_h5_dataset(grp_name, d.astype('i8'))
         return d, attr_keys
 
     def run_ant_colony(self):
